restructure and improve
initially this was only built for spirit sites which just had prepresses named by garment code. but for team shops, the prepresses come from existing production orders and are named with the design number. i added logic to locate a reference design number if present and find the prepress based on the design number. otherwise the garment code and style number will be used.

also made logical improvements for parsing netsuite data and building roster objects.